admin_b2c/requirements.txt
Flask==1.1.2
Flask-Caching==1.10.1
Flask-SQLAlchemy==2.4.4
isort==5.5.1
itsdangerous==1.1.0
Jinja2==2.11.2
mysql-connector-python==8.0.21
mysqlclient==2.0.1
redis==4.1.2
requests==2.27.1
SQLAlchemy==1.3.19
Werkzeug==1.0.1
wrapt==1.12.1
MarkupSafe==1.1.1
msal>=1.7,<2
flask-session>=0.3.2,<0.5